fix: add interface for AbstractQueue
